Rename GroupAddButtons component
